Basic changes
-changed README
-cleaned up shorten.php in the same way as the original ltfu.pw

Full disclosure, I’m a new programmer, and at this point I’m done with
all the code changes (html doesn’t count). I think we’ll need to obtain
a server to use it on to use it, since hightechlowlife.eu isn’t exactly
the basis of a good shortened address. An htll.[anything] address would
be particularly nice, and hopefully we can keep a .pw address.